+++ valid8/tests/conftest.py
@@ -1,6 +1,12 @@
+import sys
+
+import os
 from py.xml import html
 import pytest
 from setuptools_scm.git import GitWorkdir
+
+
+sys.path.append(os.path.join(os.path.dirname(__file__), 'helpers'))
 
 
 @pytest.mark.hookwrapper